Create C. Simple Calculator.cpp

+++ C++ Assignments/Codeforces/Sheet1/C. Simple Calculator.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-  int x,y;
+  long long int x,y;
   cin>>x>>y;
   cout<<x<<" + "<<y<<" = "<<x+y<<"\n";
   cout<<x<<" * "<<y<<" = "<<x*y<<"\n";
